CBIS_DDSM_Calcification/model/.mdai/requirements.txt
numpy
imageio
opencv_contrib_python==4.7.0.72
opencv_python==4.6.0.66
Pillow==9.5.0
pydicom==2.3.0
torch==1.12.1
torchvision==0.13.1
